test(e2e): consolidate review specs into one plan-detail-review.spec.ts
The Review section is a single sub-area of plan-detail, but it was split across
three files (plan-detail-review / -review-footer / -review-flow), adding a level
below plan-detail-review and violating the e2e convention (AGENTS.md: one
<feature>.spec.ts per sub-area; never split by fixture). Footer and flow are
facets of the Review section, not separate sub-areas.

Merge all three into plan-detail-review.spec.ts as describe blocks. Each block
now self-configures the project/workspace settings it needs (approval rule via
a shared setupApproval() helper, plan-check gate, ERROR review config) in its
own beforeAll, so the blocks are order-independent despite sharing the
workspace-level WORKSPACE_APPROVAL setting — the file-level beforeAll only
snapshots originals + opens the shared browser, and afterAll restores them.

Verified: 15 passed (14 review tests + setup) in one file.
